+++ yandex/base_lb_test.go
@@ -117,7 +117,11 @@ func getSubnetIPMap(instanceNames []string) (map[string][]string, error) {
 		}
 		subnetID := ifs[0].GetSubnetId()
 		address := ifs[0].GetPrimaryV4Address().GetAddress()
-		result[subnetID] = append(result[subnetID], address)
+		if _, ok := result[subnetID]; ok {
+			result[subnetID] = append(result[subnetID], address)
+		} else {
+			result[subnetID] = []string{address}
+		}
 	}
 
 	return result, nil