+++ yandex/config_test.go
@@ -117,8 +117,8 @@ func (s *userAgentMockServerAPIEndpoint) List(ctx context.Context, r *endpoint.L
 func localListener(t *testing.T) net.Listener {
 	l, err := net.Listen("tcp", "127.0.0.1:0")
 	if err != nil {
-		_, err = net.Listen("tcp6", "[::1]:0")
-		t.Fatal(err, "failed to listen on an any port")
+		l, err = net.Listen("tcp6", "[::1]:0")
+		t.Fatal(err, "failed to listen on a any port")
 	}
 	return l
 }