+++ yandex/data_source_yandex_mdb_clickhouse_cluster.go
@@ -178,6 +178,10 @@ func dataSourceYandexMDBClickHouseCluster() *schema.Resource {
 							Type:     schema.TypeBool,
 							Computed: true,
 						},
+						"serverless": {
+							Type:     schema.TypeBool,
+							Computed: true,
+						},
 					},
 				},
 			},