+++ yandex/resource_yandex_function_trigger_test.go
@@ -305,7 +305,6 @@ resource "yandex_function_trigger" "test-trigger" {
 	`, regName, devName, name)
 }
 
-//nolint:unused
 func testYandexFunctionTriggerMessageQueue(name, queueID, serviceAccountID string) string {
 	return fmt.Sprintf(`
 resource "yandex_function_trigger" "test-trigger" {