+++ yandex/resource_yandex_kubernetes_cluster.go
@@ -149,8 +149,7 @@ func resourceYandexKubernetesCluster() *schema.Resource {
 								Schema: map[string]*schema.Schema{
 									"region": {
 										Type:     schema.TypeString,
-										Optional: true,
-										Computed: true,
+										Required: true,
 										ForceNew: true,
 									},
 									"location": {