+++ yandex/resource_yandex_mdb_clickhouse_cluster.go
@@ -226,6 +226,11 @@ func resourceYandexMDBClickHouseCluster() *schema.Resource {
 							Optional: true,
 							Default:  false,
 						},
+						"serverless": {
+							Type:     schema.TypeBool,
+							Optional: true,
+							Default:  false,
+						},
 					},
 				},
 			},